move changes code into JournalChanges module
