+++ lib/plugins/acts_as_journalized/lib/journal_changes.rb
@@ -60,6 +60,8 @@ module JournalChanges
     @changes.merge!(get_association_changes(predecessor, 'customizable', 'custom_fields', :custom_field_id, :value))
   end
 
+  private
+
   def get_association_changes(predecessor, journal_association, association, key, value)
     journal_assoc_name = "#{journal_association}_journals"
 
@@ -73,7 +75,57 @@ module JournalChanges
       new_journals = send(journal_assoc_name).map(&:attributes)
       old_journals = predecessor.send(journal_assoc_name).map(&:attributes)
 
-      JournalManager.changes_on_association(new_journals, old_journals, association, key, value)
+      changes_on_association(new_journals, old_journals, association, key, value)
+    end
+  end
+
+  def changes_on_association(current, predecessor, association, key, value)
+    merged_journals = merge_reference_journals_by_id(current, predecessor, key.to_s, value.to_s)
+
+    changes = added_references(merged_journals)
+                .merge(removed_references(merged_journals))
+                .merge(changed_references(merged_journals))
+
+    to_changes_format(changes, association.to_s)
+  end
+
+  def added_references(merged_references)
+    merged_references
+      .select { |_, (old_value, new_value)| old_value.nil? && new_value.present? }
+  end
+
+  def removed_references(merged_references)
+    merged_references
+      .select { |_, (old_value, new_value)| old_value.present? && new_value.nil? }
+  end
+
+  def changed_references(merged_references)
+    merged_references
+      .select { |_, (old_value, new_value)| old_value.present? && new_value.present? && old_value.strip != new_value.strip }
+  end
+
+  def to_changes_format(references, key)
+    references.each_with_object({}) do |(id, (old_value, new_value)), result|
+      result["#{key}_#{id}"] = [old_value, new_value]
+    end
+  end
+
+  def merge_reference_journals_by_id(new_journals, old_journals, id_key, value)
+    all_associated_journal_ids = new_journals.map { |j| j[id_key] } | old_journals.map { |j| j[id_key] }
+
+    all_associated_journal_ids.each_with_object({}) do |id, result|
+      result[id] = [select_and_combine_journals(old_journals, id, id_key, value),
+                    select_and_combine_journals(new_journals, id, id_key, value)]
+    end
+  end
+
+  def select_and_combine_journals(journals, id, key, value)
+    selected_journals = journals.select { |j| j[key] == id }.map { |j| j[value] }
+
+    if selected_journals.empty?
+      nil
+    else
+      selected_journals.join(',')
     end
   end
