--- spec_legacy/unit/journal_spec.rb
@@ -1,70 +1,0 @@
-#-- encoding: UTF-8
-
-#-- copyright
-# OpenProject is an open source project management software.
-# Copyright (C) 2012-2020 the OpenProject GmbH
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License version 3.
-#
-# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
-# Copyright (C) 2006-2017 Jean-Philippe Lang
-# Copyright (C) 2010-2013 the ChiliProject Team
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-# See docs/COPYRIGHT.rdoc for more details.
-#++
-require_relative '../legacy_spec_helper'
-
-describe Journal,
-         type: :model,
-         with_settings: { notified_events: %w(work_package_added work_package_updated) } do
-  fixtures :all
-
-  specify 'creating a journal should update the updated_on value of the parent record (touch)' do
-    @user = FactoryBot.create(:user)
-    @project = FactoryBot.create(:project)
-    @issue = FactoryBot.create(:work_package, project: @project)
-    start = @issue.updated_at
-    sleep(1) # TODO: massive hack to make sure the timestamps are different. switch to timecop later
-
-    assert_difference('Journal.count') do
-      @issue.add_journal(@user, 'A note')
-      @issue.save
-    end
-
-    refute_equal start, @issue.reload.updated_at
-  end
-
-  specify 'accessing #journaled on a Journal should not error (parent class)' do
-    journal = Journal.new
-    expect {
-      assert_equal nil, journal.journable
-    }.not_to raise_error
-  end
-
-  specify 'setting journal fields through the journaled object for creation' do
-    @issue = FactoryBot.create(:work_package)
-
-    @issue.add_journal @issue.author, 'Test setting fields on Journal from Issue'
-    assert_difference('Journal.count') do
-      assert @issue.save
-    end
-
-    assert_equal 'Test setting fields on Journal from Issue', @issue.last_journal.notes
-    assert_equal @issue.author, @issue.last_journal.user
-  end
-end
